serial

+++ serial.h
@@ -10,7 +10,7 @@
 #ifndef _SERIAL_H
 #define _SERIAL_H
 
-void serial_init(uint16_t USART_BAUDRATE);
+void serial_init(uint32_t USART_BAUDRATE);
 void serial_put_char(char);
 char serial_get_char(void);
 void serial_put_str(char *);
